+++ src/components/shared/activity-card/shared/schemas.ts
