+++ src/stories/components/activity-cards/release-card.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 
-import ReleaseCard from "@/components/shared/activity-cards/release-card";
+import ReleaseCard from "@/components/shared/activity-card/release-card";
 
 const meta = {
   title: "Components/ActivityCards/Release",
